+++ src/pages/product-master/Windows/ProductMasterWindow.js
@@ -1,102 +1,44 @@
 // ** Custom Imports
-import Window from 'src/components/Shared/Window'
 import CustomTabPanel from 'src/components/Shared/CustomTabPanel'
-
-// **Tabs
-import ProductMasterTab from 'src/pages/product-master/Tabs/productMasterTab'
-import ProductDispersalTab from 'src/pages/product-master/Tabs/productDispersalTab'
-import ProductLegTab from 'src/pages/product-master/Tabs/productLegTab'
-import ProductFieldTab from 'src/pages/product-master/Tabs/productFieldTab'
-import ProductAgentTab from 'src/pages/product-master/Tabs/productAgentTab'
-import ProductCurrenciesTab from '../Tabs/productCurrenciesTab'
-import ProductCountriesTab from '../Tabs/productCountriesTab'
-import ProductSchedulesTab from '../Tabs/productSchedulesTab'
+import { useState } from 'react'
+import { CustomTabs } from 'src/components/Shared/CustomTabs'
+import ProductMasterTab from '../Tabs/productMasterTab'
 
 const ProductMasterWindow = ({
-  onClose,
-  tabs,
-  activeTab,
-  setActiveTab,
-  width,
-  height,
-  onSave,
-  productMasterValidation,
-  typeStore,
-  functionStore,
-  commissionBaseStore,
-  interfaceStore,
-  languageStore,
-  correspondentStore,
-  setCorrespondentStore,
-  lookupCorrespondent,
-
-  //countries inline edit grid
-  countriesGridValidation,
-  countriesInlineGridColumns,
-
-  //monetary inline edit grid
-  monetariesGridValidation,
-  monetariesInlineGridColumns,
-
-  //dispersals tab
-  dispersalsGridData,
-  getDispersalsGridData,
-  addProductDispersal,
-  delProductDispersal,
-  popupDispersal,
-
-  //schedules tab
-  schedulesGridValidation,
-  schedulesInlineGridColumns,
-
-  //amount ranges tab
-  productLegValidation,
-
-  //productLegsGridValidation,
-
-  //productLegsInlineGridColumns,
-  currencyStore,
-  plantStore,
-  countryStore,
-  dispersalStore,
-  scheduleRangeGridValidation,
-  scheduleRangeInlineGridColumns,
-
-  //product dispersal agents tab
-  agentsHeaderValidation,
-  agentsGridValidation,
-  agentsInlineGridColumns,
-  onDispersalSelection,
-
-  maxAccess
+  labels,
+  maxAccess,
+  height
 }) => {
+  const [editMode , setEditMode] = useState(false)
+  const [activeTab , setActiveTab] = useState(0)
+
+  const tabs = [
+    { label: labels.main },
+    { label: labels.countries, disabled: !editMode },
+    { label: labels.monetary , disabled: !editMode },
+    { label: labels.dispersal, disabled: !editMode },
+    { label: labels.schedules, disabled: !editMode },
+    { label: labels.amountRange, disabled: !editMode },
+    { label: labels.field, disabled: !editMode },
+    { label: labels.agent, disabled: !editMode }
+  ]
+
   return (
-    <Window
-      id='ProductMasterWindow'
-      Title='Product Master'
-      onClose={onClose}
-      tabs={tabs}
-      activeTab={activeTab}
-      setActiveTab={setActiveTab}
-      width={width}
-      height={height}
-      onSave={onSave}
-    >
-      <CustomTabPanel index={0} value={activeTab}>
+    <>
+        <CustomTabs  tabs={tabs} activeTab={activeTab} setActiveTab={setActiveTab} />
+
+      <CustomTabPanel height={height} index={0} value={activeTab}>
         <ProductMasterTab
-          productMasterValidation={productMasterValidation}
-          typeStore={typeStore}
-          functionStore={functionStore}
-          commissionBaseStore={commissionBaseStore}
-          interfaceStore={interfaceStore}
-          languageStore={languageStore}
           maxAccess={maxAccess}
-          correspondentStore={correspondentStore}
-          setCorrespondentStore={setCorrespondentStore}
-          lookupCorrespondent={lookupCorrespondent}
+          labels={labels}
+          setEditMode={setEditMode}
+
+          // correspondentStore={correspondentStore}
+          // setCorrespondentStore={setCorrespondentStore}
+          // lookupCorrespondent={lookupCorrespondent}
         />
       </CustomTabPanel>
-      <CustomTabPanel index={1} value={activeTab}>
+     <CustomTabPanel index={1} value={activeTab}>
         <ProductCountriesTab
           productMasterValidation={productMasterValidation}
           countriesGridValidation={countriesGridValidation}
@@ -104,7 +46,7 @@ const ProductMasterWindow = ({
           maxAccess={maxAccess}
         />
       </CustomTabPanel>
-      <CustomTabPanel index={2} value={activeTab}>
+      {/*  <CustomTabPanel index={2} value={activeTab}>
         <ProductCurrenciesTab
           productMasterValidation={productMasterValidation}
           monetariesGridValidation={monetariesGridValidation}
@@ -143,15 +85,15 @@ const ProductMasterWindow = ({
           dispersalStore={dispersalStore}
           maxAccess={maxAccess}
         />
-      </CustomTabPanel>
-      <CustomTabPanel index={6} value={activeTab}>
+      </CustomTabPanel> */}
+      {/* <CustomTabPanel index={6} value={activeTab}> */}
         {/* <ProductFieldTab
           productFieldGridData={productFieldGridData}
           dispersalStore={dispersalStore}
           maxAccess={maxAccess}
         /> */}
-      </CustomTabPanel>
-      <CustomTabPanel index={7} value={activeTab}>
+      {/* </CustomTabPanel> */}
+      {/* <CustomTabPanel index={7} value={activeTab}>
         <ProductAgentTab
           onDispersalSelection={onDispersalSelection}
           dispersalsGridData={dispersalsGridData}
@@ -161,8 +103,8 @@ const ProductMasterWindow = ({
           dispersalStore={dispersalStore}
           maxAccess={maxAccess}
         />
-      </CustomTabPanel>
-    </Window>
+      </CustomTabPanel> */}
+    </>
   )
 }
